Refactoring scripts
