+++ Assets/AutoTranslate/Scripts/GoogleApiFree.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d9d03995c8169154e8d6a8958fcdb03c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
