+++ Assets/AutoTranslate/Scripts/SimpleLocalization/LocalizeDropdownEvent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e4bfe6b2382c2464c91cc16b81385079
+guid: f530283b664e5a846a2d33f7235b3797
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
